requirements.txt
pandas==1.0.1
matplotlib==3.1.3
scipy==1.4.1
streamlit==0.56.0
numpy==1.18.1
s3fs==0.4.0
Pillow==7.0.0
